requirements.txt
Django==3.2.8
djangorestframework==3.12.4
mysqlclient==2.1.0
django-cors-headers==3.10.1
